add do_other_test(), a simple way to run a test in the 'other' suite
