+++ tests/other/malloc_implicit/test.out
@@ -1,0 +1,2 @@
+ok
+
